Mostly gui rework
Merged Inventory and stats window into one window
renamed inventory window to random numbers
moved font to font folder, also reworked and merged some textures together.

+++ Project/Text/FloatingDamage.hpp
@@ -9,7 +9,7 @@ private:
 
 public:
 	FloatingDamage();
-	FloatingDamage(std::string iconName,sf::Font& font,double duration = 3,int len = -1,int offsetX = 0,int offsetY = 0);
+	FloatingDamage(std::string iconName,sf::Font& font,double duration = 3,int len = -1,double offsetX = 0,double offsetY = 0);
 	FloatingDamage(FloatingDamage &fdmg);
 
 	FloatingDamage& operator=(int damageDone);
